Adds medical profile retrieval by ID
Enables retrieval of medical profiles by their unique ID.

This is done by adding a new query to the MedicalProfileDao
and a corresponding function to the MedicalProfileRepository.

Updates MedicalProfileEditorViewModel to use the application instance for database access.

Also, refactors the IMPACT_THRESHOLD constant to reside only once inside the services.

Fixes an issue with the WifiDirect service not properly receiving events.